+++ tests/test_utils.py
@@ -5,13 +5,25 @@ from django.contrib.contenttypes.models import ContentType
 from django.test import TestCase
 
 from cms.models import Page, User
+from cms.test_utils.testcases import CMSTestCase
+from cms.toolbar.utils import get_object_edit_url, get_object_preview_url
+
+from djangocms_versioning.constants import (
+    ARCHIVED,
+    DRAFT,
+    PUBLISHED,
+    UNPUBLISHED,
+)
 
 from djangocms_navigation.models import MenuContent
+from djangocms_navigation.test_utils import factories
 from djangocms_navigation.test_utils.app_1.models import TestModel1, TestModel2
 from djangocms_navigation.test_utils.app_2.models import TestModel3, TestModel4
 from djangocms_navigation.test_utils.polls.models import PollContent
 from djangocms_navigation.utils import (
+    get_latest_page_content_for_page_grouper,
     is_model_supported,
+    is_preview_or_edit_mode,
     supported_content_type_pks,
     supported_models,
 )
@@ -84,3 +96,84 @@ class IsModelSupportedTestCase(TestCase):
         unexpected_content_types = ContentType.objects.get_for_models(User)
         for model in unexpected_content_types:
             self.assertFalse(is_model_supported(MenuContent, model))
+
+
+class IsPreviewEditModeTestCase(CMSTestCase):
+    """
+    Test case for the utility: is_preview_or_edit_mode
+    """
+    def setUp(self):
+        self.page_content = factories.PageContentWithVersionFactory(
+            version__created_by=self.get_superuser(),
+            version__state=PUBLISHED,
+        )
+
+    def test_live_endpoint(self):
+        live_endpoint = self.page_content.get_absolute_url(language=self.page_content.language)
+
+        response = self.client.get(live_endpoint)
+        actual = is_preview_or_edit_mode(response.wsgi_request)
+
+        self.assertEqual(response.status_code, 200)
+        self.assertFalse(actual)
+
+    def test_preview_endpoint(self):
+        preview_endpoint = get_object_preview_url(self.page_content, language=self.page_content.language)
+
+        with self.login_user_context(self.get_superuser()):
+            response = self.client.get(preview_endpoint)
+        actual = is_preview_or_edit_mode(response.wsgi_request)
+
+        self.assertEqual(response.status_code, 200)
+        self.assertTrue(actual)
+
+    def test_edit_endpoint(self):
+        edit_endpoint = get_object_edit_url(self.page_content, language=self.page_content.language)
+
+        with self.login_user_context(self.get_superuser()):
+            response = self.client.get(edit_endpoint)
+        actual = is_preview_or_edit_mode(response.wsgi_request)
+
+        self.assertEqual(response.status_code, 200)
+        self.assertTrue(actual)
+
+
+class PreviewEditModeTestCase(TestCase):
+    """
+    Test case for the utility: get_latest_page_content_for_page_grouper
+    """
+    def test_draft_page_is_returned(self):
+        page_content = factories.PageContentWithVersionFactory(
+            version__state=DRAFT,
+        )
+        actual = get_latest_page_content_for_page_grouper(page_content.page, page_content.language)
+
+        self.assertEqual(page_content, actual)
+        self.assertIsNotNone(actual)
+
+    def test_published_page_is_returned(self):
+        page_content = factories.PageContentWithVersionFactory(
+            version__state=PUBLISHED,
+        )
+        actual = get_latest_page_content_for_page_grouper(page_content.page, page_content.language)
+
+        self.assertEqual(page_content, actual)
+        self.assertIsNotNone(actual)
+
+    def test_archived_page_is_not_returned(self):
+        page_content = factories.PageContentWithVersionFactory(
+            version__state=ARCHIVED,
+        )
+        actual = get_latest_page_content_for_page_grouper(page_content.page, page_content.language)
+
+        self.assertNotEqual(page_content, actual)
+        self.assertIsNone(actual)
+
+    def test_unpublished_page_is_not_returned(self):
+        page_content = factories.PageContentWithVersionFactory(
+            version__state=UNPUBLISHED,
+        )
+        actual = get_latest_page_content_for_page_grouper(page_content.page, page_content.language)
+
+        self.assertNotEqual(page_content, actual)
+        self.assertIsNone(actual)